Fix DataModels and create basic get all for each service

+++ Fantasy Freaks/Models/DefenseDataModel.cs
@@ -14,50 +14,50 @@ namespace Fantasy_Freaks.Models
         [Column("Team_Name")]
         public string TeamName { get; set; }
         [Column("PtsAllow")]
-        public int PointsAllowed { get; set; }
+        public double PointsAllowed { get; set; }
         [Column("TotYdsAllow")]
-        public int TotalYardsAllowed { get; set; }
+        public double TotalYardsAllowed { get; set; }
         [Column("OffPlayAllow")]
-        public int OffPlaysAllowed { get; set; }
+        public double OffPlaysAllowed { get; set; }
         [Column("YdsPerPlay")]
-        public int YardsPerPlay { get; set; }
+        public double YardsPerPlay { get; set; }
         [Column("Turnovers")]
-        public int Turnovers { get; set; }
+        public double Turnovers { get; set; }
         [Column("ForceFumb")]
-        public int ForcedFumbles { get; set; }
+        public double ForcedFumbles { get; set; }
         [Column("TotFirstDownAllow")]
-        public int TotalFirstDownsAllowed { get; set; }
-        public int Completions { get; set; }
+        public double TotalFirstDownsAllowed { get; set; }
+        public double Completions { get; set; }
         [Column("PassAtmptAllow")]
-        public int PassAttemptsAllowed { get; set; }
+        public double PassAttemptsAllowed { get; set; }
         [Column("PassYdsAllow")]
-        public int PassingYardsAllowed { get; set; }
+        public double PassingYardsAllowed { get; set; }
         [Column("PassTDAllow")]
-        public int PassingTouchdownsAllowed { get; set; }
-        public int Interceptions { get; set; }
+        public double PassingTouchdownsAllowed { get; set; }
+        public double Interceptions { get; set; }
         [Column("NetYdsAllowPerPass")]
-        public int NetYardsAllowedPerPass { get; set; }
+        public double NetYardsAllowedPerPass { get; set; }
         [Column("PassFirstDownAllow")]
-        public int FirstDownPassesAllowed { get; set; }
+        public double FirstDownPassesAllowed { get; set; }
         [Column("RushAtmptAllow")]
-        public int RushintAttemptsAllowed { get; set; }
+        public double RushintAttemptsAllowed { get; set; }
         [Column("RushYdsAllow")]
-        public int RushingYardsAllowed { get; set; }
+        public double RushingYardsAllowed { get; set; }
         [Column("RushTDAllow")]
-        public int RushingTouchdownsAllowed { get; set; }
+        public double RushingTouchdownsAllowed { get; set; }
         [Column("YdsPerRushAllow")]
-        public int YardsPerRushAllowed { get; set; }
+        public double YardsPerRushAllowed { get; set; }
         [Column("RushFirstDownAllow")]
-        public int RushingFirstDownsAllowed { get; set; }
+        public double RushingFirstDownsAllowed { get; set; }
         [Column("PenCommited")]
-        public int PenaltiesCommitted { get; set; }
+        public double PenaltiesCommitted { get; set; }
         [Column("PenYdsAllow")]
-        public int PenaltyYardsAllowed { get; set; }
+        public double PenaltyYardsAllowed { get; set; }
         [Column("FirstDownViaPen")]
-        public int FirstDownsViaPenalties { get; set; }
+        public double FirstDownsViaPenalties { get; set; }
         [Column("PercDrivesEndOffScore")]
-        public int PercDrivesEndOffScore { get; set; }
+        public double PercDrivesEndOffScore { get; set; }
         [Column("PercDrivesEndTO")]
-        public int PercDrivesEndTO { get; set; }
+        public double PercDrivesEndTO { get; set; }
     }
 }
